backend-work/app/requirements.txt
fastapi
uvicorn
google-generativeai
scikit-learn
numpy
pandas
pyttsx3
python-dotenv
pydantic